refactor[#287] : 컨펌 모달 제거
