+++ src/features/setting/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from "./usePasswordCheckModal.tsx";
-export * from "./usePasswordChangeModal.tsx";
 export * from "./useSettingPermission.tsx";
-export * from "./usePasswordSetModal";
