+++ src/pages/my-page/setting/manage-account/page.tsx
@@ -1,9 +1,8 @@
-import { usePasswordChangeModal } from "@/features/setting/hooks";
-import {
-  useSettingPermission,
-  usePasswordSetModal,
-} from "@/features/setting/hooks";
+import { useQueryClient } from "@tanstack/react-query";
+import { useSettingPermission } from "@/features/setting/hooks";
 import { AccountCancellationModal } from "@/features/setting/ui";
+import { PasswordChangeModal } from "@/features/setting/ui";
+import { PasswordSetModal } from "@/features/setting/ui";
 import type { MyInfo } from "@/entities/auth/api";
 import { useGetMyInfo } from "@/entities/auth/api";
 import { SOCIAL_TYPE } from "@/entities/auth/constants";
@@ -57,10 +56,27 @@ const AccountEmail = ({
 };
 
 const PasswordChangeButton = () => {
-  const handleOpenPasswordChangeModal = usePasswordChangeModal();
+  const queryClient = useQueryClient();
+  const { handleOpen, onClose } = useModal(
+    () => <PasswordChangeModal onClose={onClose} />,
+    {
+      beforeClose: () => {
+        /**
+         * 만약 mutation 이 진행 중이라면 모달을 닫는 행위를
+         * 중지 시킵니다.
+         */
+        const mutationCache = queryClient.getMutationCache().find({
+          mutationKey: ["putChangePassword"],
+        });
+        if (mutationCache?.state.status === "pending") {
+          return true;
+        }
+      },
+    },
+  );
 
   return (
-    <button className="setting-item" onClick={handleOpenPasswordChangeModal}>
+    <button className="setting-item" onClick={handleOpen}>
       <p>비밀번호 변경</p>
       <div className="flex items-center text-grey-500">
         <span className="body-2">●●●●●●●●</span>
@@ -72,10 +88,28 @@ const PasswordChangeButton = () => {
 
 const PasswordSetButton = () => {
   // TODO 비밀번호 설정 모달로 변경 하기
-  const handleOpenPasswordSetModal = usePasswordSetModal();
+  const queryClient = useQueryClient();
+
+  const { handleOpen, onClose } = useModal(
+    () => <PasswordSetModal onClose={onClose} />,
+    {
+      beforeClose: () => {
+        /**
+         * 만약 mutation 이 진행 중이라면 모달을 닫는 행위를
+         * 중지 시킵니다.
+         */
+        const mutationCache = queryClient.getMutationCache().find({
+          mutationKey: ["putSetPassword"],
+        });
+        if (mutationCache?.state.status === "pending") {
+          return true;
+        }
+      },
+    },
+  );
 
   return (
-    <button className="setting-item" onClick={handleOpenPasswordSetModal}>
+    <button className="setting-item" onClick={handleOpen}>
       <p>비밀번호 설정</p>
       <div className="flex items-center text-grey-500">
         <span className="body-2">비밀번호를 설정해 주세요</span>
